+++ stable_baselines3/dqn/policies.py
@@ -90,7 +90,7 @@ class DQNPolicy(BasePolicy):
     :param activation_fn: Activation function
     :param features_extractor_class: Features extractor to use.
     :param features_extractor_kwargs: Keyword arguments
-        to pass to the feature extractor.
+        to pass to the features extractor.
     :param normalize_images: Whether to normalize images or not,
          dividing by 255.0 (True by default)
     :param optimizer_class: The optimizer to use,
@@ -158,10 +158,9 @@ class DQNPolicy(BasePolicy):
         self.optimizer = self.optimizer_class(self.parameters(), lr=lr_schedule(1), **self.optimizer_kwargs)
 
     def make_q_net(self) -> QNetwork:
-        # Make sure we always have separate networks for feature extractors etc
-        features_extractor = self.features_extractor_class(self.observation_space, **self.features_extractor_kwargs)
-        features_dim = features_extractor.features_dim
-        return QNetwork(features_extractor=features_extractor, features_dim=features_dim, **self.net_args).to(self.device)
+        # Make sure we always have separate networks for features extractors etc
+        net_args = self._update_features_extractor(self.net_args, features_extractor=None)
+        return QNetwork(**net_args).to(self.device)
 
     def forward(self, obs: th.Tensor, deterministic: bool = True) -> th.Tensor:
         return self._predict(obs, deterministic=deterministic)